Update modal
PersonalMovieModal no longer relies on bootstrap. I wanted to be able to build the modal functionality from the ground up. Trailer will be next and this will allow the user to have a more streamlined experience. A modal background has also been added. Clicking on the background will close the movie modal - one of my primary reasons for wanting to move away from the React Bootstrap modal. So excited to have made this commit